Refactor downloads by year
* replace multiple buttons with a dropdown by year
* update forms table
* update gems
