+++ app/javascript/application.js
@@ -1,4 +1,4 @@
 // Configure your import map in config/importmap.rb. Read more: https://github.com/rails/importmap-rails
-//= require controllers
+import "controllers"
 import "@hotwired/turbo-rails"
 Turbo.session.drive = false
